refactor: shrink PostingEntry to 6 bytes (u32+u8+u8)
Adopt trigrep's compact format: 6 bytes per posting entry instead of 16.
This reduces index size by 2.7x and improves cache behavior.

Changes:
- loc_mask: u64 -> u8, using offset % 8 (enables proper rotate-and-AND
  adjacency checks since consecutive offsets map to adjacent bits)
- next_mask: u32 -> u8, using multiplicative bloom hash (0x9E) mapped
  to 3 bits for 8 Bloom positions
- check_adjacency: now uses u8::rotate_left(1) for correct adjacency
- bloom_hash: exposed as public function for query-time checks

+++ tgrep-core/src/builder.rs
@@ -134,8 +134,8 @@ pub fn write_index_from_snapshot(
         for &fid in posting_list {
             let entry = PostingEntry {
                 file_id: fid,
-                loc_mask: u64::MAX, // all bits set = no filtering
-                next_mask: u32::MAX,
+                loc_mask: u8::MAX, // all bits set = no filtering
+                next_mask: u8::MAX,
             };
             postings_file.write_all(&entry.encode())?;
         }
